requirements.txt
cx-Oracle == 8.3.0
robotframework >= 3.2.2
robotframework-pythonlibcore >= 2.2.1
